+++ src/main/java/com/back/global/init/TestInitData.kt
@@ -6,6 +6,7 @@ import com.back.domain.news.fake.dto.FakeNewsDto
 import com.back.domain.news.fake.service.FakeNewsService
 import com.back.domain.news.real.dto.RealNewsDto
 import com.back.domain.news.real.service.NewsDataService
+import com.back.domain.news.today.service.TodayNewsService
 import com.back.domain.quiz.daily.service.DailyQuizService
 import com.back.domain.quiz.detail.dto.DetailQuizDto
 import com.back.domain.quiz.detail.entity.Option
@@ -34,6 +35,9 @@ class TestInitData(
     private val factQuizService: FactQuizService,
     private val dailyQuizService: DailyQuizService
 ) {
+    @Autowired
+    private lateinit var todayNewsService: TodayNewsService
+
     @Autowired
     @Lazy
     private val self: TestInitData? = null
@@ -195,7 +199,7 @@ class TestInitData(
         val savedNewsList = newsDataService.saveAllRealNews(newsList)
 
         if (!savedNewsList.isEmpty()) {
-            newsDataService.setTodayNews(savedNewsList[1].id)
+            todayNewsService.setTodayNews(savedNewsList[1].id)
         }
     }
